First version working with simulator

+++ src/tools.cpp
@@ -30,12 +30,12 @@ VectorXd Tools::CalculateRMSE(const vector<VectorXd> &estimations,
         VectorXd residual = estimations[i] - ground_truth[i];
         
         //coefficient-wise multiplication
-        residual = residual.array()*residual.array();
+        residual = residual.array() * residual.array();
         rmse += residual;
     }
     
     //calculate the mean
-    rmse = rmse/estimations.size();
+    rmse = rmse / estimations.size();
     
     //calculate the squared root
     rmse = rmse.array().sqrt();
@@ -47,18 +47,18 @@ VectorXd Tools::CalculateRMSE(const vector<VectorXd> &estimations,
 MatrixXd Tools::CalculateJacobian(const VectorXd& x_state) {
     MatrixXd Hj(3,4);
     //recover state parameters
-    float px = x_state(0);
-    float py = x_state(1);
-    float vx = x_state(2);
-    float vy = x_state(3);
+    double px = x_state(0);
+    double py = x_state(1);
+    double vx = x_state(2);
+    double vy = x_state(3);
     
     //pre-compute a set of terms to avoid repeated calculation
-    float c1 = px*px+py*py;
-    float c2 = sqrt(c1);
-    float c3 = (c1*c2);
+    double c1 = px*px + py*py;
+    double c2 = sqrt(c1);
+    double c3 = c1*c2;
     
     //check division by zero
-    if(fabs(c1) < 0.0001){
+    if(fabs(c1) < 0.0001) {
         cout << "CalculateJacobian () - Error - Division by Zero" << endl;
         return Hj;
     }
